Notebooks/demos/foodvision_mini/requirements.txt
gradio==3.41.2
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
